add quantity

+++ src/app/business.InterfaceLayer/store/cartStore.ts
@@ -1,34 +1,75 @@
 "use client";
 
-import { useState, useEffect } from "react";
 import { create } from "zustand";
-import { Products } from "../types";
+import { Items, Products } from "../types";
 
 interface CartStore {
-  items: Products[];
+  items: Items[];
   addToCart: (product: Products) => void;
   removeFromCart: (productId: number) => void;
+  removeItem: (productId: number) => void;
   clearCart: () => void;
 }
 
 export const useCartStore = create<CartStore>((set) => {
-  // Получаем данные из localStorage при инициализации состояния
-  const initialItems = JSON.parse(localStorage.getItem("cartItems") || "[]");
+  const initialItems: Items[] = JSON.parse(localStorage.getItem("cartItems") || "[]");
+
+  const calculateTotal = (items: Items[]): number => {
+    return items.reduce((acc, item) => acc + item.sumProduct, 0);
+  };
 
   return {
     items: initialItems,
     addToCart: (product) =>
       set((state) => {
-        const updatedItems = [...state.items, product];
-        localStorage.setItem("cartItems", JSON.stringify(updatedItems)); // Обновляем localStorage
-        return { items: updatedItems };
+        const existingItemIndex = state.items.findIndex((item) => item.id === product.id);
+        if (existingItemIndex !== -1) {
+          const updatedItems = [...state.items];
+          updatedItems[existingItemIndex].quantity += 1;
+          updatedItems[existingItemIndex].sumProduct += product.actualPrice;
+          const total = calculateTotal(updatedItems);
+          localStorage.setItem("cartItems", JSON.stringify(updatedItems));
+          return { items: updatedItems, total };
+        } else {
+          const newItems = [
+            ...state.items,
+            { ...product, quantity: 1, sumProduct: product.actualPrice },
+          ];
+          const total = calculateTotal(newItems);
+          localStorage.setItem("cartItems", JSON.stringify(newItems));
+          return { items: newItems, total };
+        }
+      }),
+    removeItem: (productId) =>
+      set((state) => {
+        const existingItemIndex = state.items.findIndex((item) => item.id === productId);
+        if (existingItemIndex !== -1) {
+          const updatedItems = [...state.items];
+          if (updatedItems[existingItemIndex].quantity > 1) {
+            updatedItems[existingItemIndex].quantity -= 1;
+            updatedItems[existingItemIndex].sumProduct -=
+              updatedItems[existingItemIndex].actualPrice;
+          } else {
+            updatedItems.splice(existingItemIndex, 1);
+          }
+          const total = calculateTotal(updatedItems);
+          localStorage.setItem("cartItems", JSON.stringify(updatedItems));
+          return { items: updatedItems, total };
+        } else {
+          return { items: state.items, total: state.total };
+        }
       }),
     removeFromCart: (productId) =>
       set((state) => {
         const updatedItems = state.items.filter((item) => item.id !== productId);
-        localStorage.setItem("cartItems", JSON.stringify(updatedItems)); // Обновляем localStorage
-        return { items: updatedItems };
+        const total = calculateTotal(updatedItems);
+        localStorage.setItem("cartItems", JSON.stringify(updatedItems));
+        return { items: updatedItems, total };
       }),
-    clearCart: () => set({ items: [] }),
+    clearCart: () =>
+      set(() => {
+        localStorage.removeItem("cartItems");
+        return { items: [], total: 0 };
+      }),
   };
 });
